Created apps directory, moved Packet Analyzer into that directory. Created a solution including TibiaAPI and the apps. (Developers will probably want to make their own however, to include projects not under version control.) Fixed a small compile error in the proxy.

+++ tibiaapi/Util/Proxy.cs
@@ -288,7 +288,7 @@ namespace Tibia.Util
         {
             int bytesRead = netStreamClient.EndRead(ar);
 
-            if (GetPacketType(dataClient) == PacketType.Logout)
+            if (GetPacketType(dataClient) == (byte)PacketType.Logout)
             {
                 Stop();
                 Restart();
